+++ server/templates/dashboard.html
